refactor(scan): split in-memory scan owner
